requirements.txt
python==3.7
mysql-connector-python
sshtunnel
flask
gym
opencv-python
Flask==1.1.2
flask_restplus==0.13.0
Flask_Script==2.0.6
Flask-SQLAlchemy==2.5.1
mysqlclient==2.0.3
python-dateutil==2.8.1
python-dotenv==0.13.0
Werkzeug==0.16.0
